pyproject.toml
[project]
name = "oanda-rates"
version = "0.1.0"
description = "A desktop application to fetch, store, and display OANDA financing rates."
authors = [
    { name = "skum", email = "your.email@example.com" },
]
requires-python = ">=3.11"
keywords = ["oanda", "financing-rates", "forex", "trading", "desktop-app", "pyqt6"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "requests",
    "SQLAlchemy",
    "APScheduler",
    "PyQt6",
    "matplotlib",
    "pandas",
    "pytz",
    "PyYAML",
    "pytest-qt",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "black",
    "mypy",
    "pandas-stubs",
    "types-requests",
    "pyqt6-stubs",
    "bandit",
    "radon",
    "types-PyYAML",
    "types-PyQt6",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "C", "N", "D", "B", "A", "SIM", "UP", "PL"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "src.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "apscheduler.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["src"]
